Improve JUnitXmlTestRunResultProcessor's URL generation
The processor that generates JUnit-compatible XML reports attempts to generate an URL to the specifications that are under test.

Input used for these URLs should be properly escaped.

Links should be generated not only for XEP-specifications, but also for RFCs.

If a link can't be created, prevent the report generation from crashing. Instead, just generate a report without a link.

fixes #21