Fix privilege in MUC mucTestAdminMemberListItemCheck
This test asserts that items in a member list have required fields. To do so, the test retrieves a member list.

The user that is used to retreive the member list is a non-privileged user. The XEP does not mandate that such a user must be allowed to retrieve the memberlist. To prevent issues, a privileged user should be used instead. All other tests in this class do so - it appears that this test erroneously omits the grant of that privilege in the test fixture setup. This is corrected in this commit.