Fix the actor in several tests that assert admins modifying moderator lists
In tests that verify that a user that is an admin can or cannot perform certain actions, the actor should be an admin.

Prior to this commit, the actor that was used for certain tests was an owner (a super-class of 'admin'). This is done to work around a limitation in which only three test users can be done. In certain tests, using an owner negates the purpose of the test (as an owner is allowed to do some things that an admin is not).