Fix VCard error equality check
When requesting a vCard from a non-existing user, an error must be returned.
When requesting a vCard from a user that has no vCard, an error must also be returned.

There are two tests in our code that check the above.

There is a third definition (and we have test for that), that states that both errors must be the same.

When one of the first two tests fail, running the third test is not very useful. This commit skips that test in that case.