Leave system in predicatable state after SubscriptionTests
When tests have finished executing, they should leave the state of the XMPP users in a consistent state. This prevents tests from having results that are affected by the exection of other tests.

The goal of this commit is to ensure that any change to subscription state that is made in the SubscriptionTests is undone.