CI: Switch to using an daily build of Openfire
While creating our integration tests, we frequently find bugs in Openfire (which are then quickly resolved, obviously!)

The continuous integration of our tests often fails for tests that identify a bug in Openfire, that has recently been fixed.

To reduce the impact of this problem, CI should use a (very) recent build of Openfire, instead of a fixed version.

Note that the implementation introduced here is not ideal:
- To prevent an excessive amount of downloads, the build caches an Openfire download based on the date that the download was attempted
- The script tries to download todays daily build, but goes back in date (up to 30 days) to find the last available build
- Depending on the time that a daily build is generated and the first execution of this script, the build that is used may very well be an older build than today's.

This can probably be improved upon, but that's beyond  my scripting foo.