Increase coverage of XEP-0045 to version 1.35.1
XEP-0045 version 1.35.1 was recently released by the XSF. This includes some improvements that we had suggested.

This commit updates the MUC tests to that version, which includes the bumping of the version identifier, modifications of modified texts, and the uncommenting of tests that we had already implemented in anticipation of this update.