XEP-0133: All tests should have a specification quote
It would be best all tests to have a 'quote' (that describes what specific definition is subject of the test), and for all assertions to have a message (that is shown when the assertion fails). This is an example of a failing test that doesn't have such a quote, and not a message in the failed assertion either:

```
💀 The following 1 tests failed! 💀
• XEP-0133: Service Administration, Section 4.5
      "null"
  Failure reason  : expected: <true> but was: <false>
  Stanza log file : ./target/logs/AdHocCommandIntegrationTest.testEndUserSession (Normal).log
  Test class      : class org.igniterealtime.smack.inttest.xep0133.AdHocCommandIntegrationTest
  Test method     : testEndUserSession
  ```

It's next to impossible for anyone to explain why this test is failing.

By adding a 'quote' text (in the method's annotation) and messages to assertions, we could get something like this instead:

```
💀 The following 1 tests failed! 💀
• XEP-0133: Service Administration, Section 4.5
      "An administrator may need to terminate one or all of the user's current sessions [...] if the JID is of the form <user@host/resource>, the service MUST end only the session associated with that resource.s"
  Failure reason  : Timeout after 5000ms: Expected the connection of 'endsessiontest6lmmy@example.org/93bf4e24-7e39-492d-b564-222c6bf2af0f' to be disconnected after 'admin@example.org/5067dfd9-5534-4408-9126-3aa6741b28e7' invoked the http://jabber.org/protocol/admin#end-user-session ad-hoc command using the target's full JID (but the connection remains connected).
  Stanza log file : ./target/logs/AdHocCommandIntegrationTest.testEndUserSession (Normal).log
  Test class      : class org.igniterealtime.smack.inttest.xep0133.AdHocCommandIntegrationTest
  Test method     : testEndUserSession
```

That's much more descriptive. With the XMPP stanza logs, people can use this to determine how the test failed.