XEP-0133: All assertions should have message (also includes functional changes)
This commit adds a couple of things at once (it was to time-consuming to split this up in multiple commits, sorry).

Various tests have been adjusted, mostly to depend less on behavior that is likely Openfire-specific.

Separately, all assertions now have messages associated to them. As with the previous commit, this makes it a lot easier for end-users to deduce what's going wrong when a test fails.