XEP-0133: Prefer random string over `testRunId`
Various bits in the test code use a random value. When randomness is needed, the `testRunId` is not ideal, as that's the same value during the entire test run. Although I don't think it was causing an issue in the existing implementation, it would be very easy to copy/paste a username that contained that value between test implementations, and accidentally used the exact same value when a distinct value was intended.

This commit replaces `testRunId` with randomly generted values.