Add tests for RFC 6121 Section 8.5.2.1.2 'Server Rules for Processing XML Stanzas'
Integration tests that verify that behavior defined in section 8.5.2.1.2 "Local User / localpart@domainpart / Available or Connected Resources / Presence" of section 8 "Server Rules for Processing XML Stanzas" of RFC6121.

The tests here only partially cover this paragraph, as the paragraph references other paragraphs for most of its definition. The implementation added in this commit tests the processing of presence stanzas with no type or of type 'unavailable' only. All other types are/should be subject of tests under other sections (Section 3 and Section 4.3), to which section 8.5.2.1.2 references (and which have a much more detailed definition of the desired functionality).