Add tests for RFC 6121 Section 8.5.3.2.2 'Server Rules for Processing XML Stanzas'
Integration tests that verify that behavior defined in section 8.5.3.2.2 "Local User / localpart@domainpart/resourcepart / No Resource Matches / Presence" of section 8 "Server Rules for Processing XML Stanzas" of RFC6121.

Note: this _does not_ tests the processing of presence stanzas with of type 'subscribe' or 'probe'. Those types are subject of tests under other sections (Section 3 and Section 4.3), to which section 8.5.3.2.2 references (and which have a much more detailed definition of the desired functionality).