Tests should not assume a default role when joining a MUC
When joining a MUC, it cannot be assumed that every server hands out the same default role. For example, ejabberd will, by default, make new occupants 'participants' in moderated rooms, while Openfire does not.

This commit adjusts tests that depend on having a specific role in the test fixture. They now explicitly check if the role on join matches the desired role (and applies a correction if that's not the case).

fixes #95