Update Smack from 4.5.0-beta5 to 4.5.0-beta6
This updates the reference to the Smack dependencies and fixes some issues that resulted from API changes.

Notably, the custom logger that was used in our project is no longer compatible with the Smack API. Smack now provides a logger that performs similar functionality.

In this commit, it was opted to drop the custom logger. It is encouraged that the Smack-provided logger is used instead.

This is done by changing the `sinttest.debugger` argument of the invocation of the application, from:

```
-Dsinttest.debugger=org.igniterealtime.smack.inttest.util.FileLoggerFactory
-DlogDir=logs
```

into:

```
-Dsinttest.debugger=standard,dir=./logs,console=off
```

The old 'logDir' property is no longer used.

Note that Smack's new logger _requires_ that the directory does not preexist, and can/must be created by Smack (this hopefully changes in future versions of Smack).

fixes #108