Don't wrap test exceptions in RuntimeException
Some tests (and teardown methods) catch exceptions thrown, wrap that in a RuntimeException, and rethrow that.

This is likely the result of some kind of default IDE-based code template.

Rethrowing as a RuntimeException is undesirable. It doesn't notably change the semantic of the failure. It can cause the execution of all following tests to be aborted (as SmackIntegrationTestFramework isn't expecting instances of RuntimeException).

This commit throws the exception that was previously wrapped.

fixes #99