Improve concurrency when dealing with Roster
I suspect that Smack asynchronously processes roster updates, which results in infrequent failures to obtain a roster item through Roster.getEntry() for a target that was known to have been added.

In this commit, specific bits of code listen for an event emitted by the Roster instance that is expected to be updated, rather than for a stanza that arrives (that implies that the roster is updated).

This is expected to reduce occasional failure of the affected tests.

fixes #104