Upgrade to Smack-4.5.0-beta6: Use temporary logger
With the beta6 release of Smack, we wanted to swich from our proprietary logger (that generates files with XMPP traffic dump) to one that is provided by Smack (since this release). However, this logger has an annoying bug: the directory in which logs are to be stored cannot exist prior to execution. This is generally annoying, but somewhat of a showstopper for container-based runners (which typically map a volume to be able to extract those log files from the container). A fix for this issue is proposed in https://github.com/igniterealtime/Smack/pull/656

As this bug is a blocker, a new, proprietary logger is introduced in this commit. It is pretty much a copy of the Smack-provided logger, with the issue mentioned above fixed.

The process of writing a new, proprietary logger has exposed another issue with the new Smack framework (addressed in https://github.com/igniterealtime/Smack/pull/657 ) which causes the logger-config to not be parsed properly. As a work-around, part of the config is hard-coded, and part is based on a separate system property (which conveniently is the same property as the one that was used by our first proprietary logger: `logDir`.

This all is becoming a bit of a mess. Hopefully, we can revert this commit after upgrading to a later release of Smack in which the issues that are being worked around here are fixed.