Improve message error response detection
Sadly, we can't depend on much characteristics to determine if a stanza is an error response to an earlier message stanza:
- IDs aren't required to be copied
- the body text isn't always copied (looking at Prosody)

This commit relaxes how those error stanzas are detected to just being a stanza of type error. That's rather loose, which may trip false positivies. It no longer fails to identify errors that are received.

fixes #120

Additionally, some surplus brackets were removed when used in lambdas.